Add Templates for nftables firewalling
This adds templates for nftables firewall configuration.  Obviously
if we go changing ports with this we need to make sure we can still
get to them after the fact.  This begins the process of keeping
track of such things.  I suspect this code will get more complicated
as more firewalls should get detected and supported.

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>
Signed-off-by: John Hawley <jhawley@vmware.com>