Rebuild the missing facts since they are likely missing at this point
So in order for this to run "well" the facts get ignored until the
port has been sorted out.  The obvious reason is, fact gathering
before we've sorted out the ssh port won't work.  So rebuilding
the facts after we've got it figured out.  This will obviously help
sort out the firewall rules going forward as well

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>
Signed-off-by: John Hawley <jhawley@vmware.com>